// src/types/openapi_spec.ts
export interface OpenAPIObject extends ISpecificationExtension {
  openapi: string
  info: InfoObject
  servers?: ServerObject[]
  paths: PathsObject
  components?: ComponentsObject
  security?: SecurityRequirementObject[]
  tags?: TagObject[]
  externalDocs?: ExternalDocumentationObject
  webhooks?: PathsObject
}
export interface InfoObject extends ISpecificationExtension {
  title: string
  description?: string
  termsOfService?: string
  contact?: ContactObject
  license?: LicenseObject
  version: string
}
export interface ContactObject extends ISpecificationExtension {
  name?: string
  url?: string
  email?: string
}
export interface LicenseObject extends ISpecificationExtension {
  name: string
  url?: string
}
export interface ServerObject extends ISpecificationExtension {
  url: string
  description?: string
  variables?: { [v: string]: ServerVariableObject }
}
export interface ServerVariableObject extends ISpecificationExtension {
  enum?: string[] | boolean[] | number[]
  default: string | boolean | number
  description?: string
}
export interface ComponentsObject extends ISpecificationExtension {
  schemas?: { [schema: string]: SchemaObject | ReferenceObject }
  responses?: { [response: string]: ResponseObject | ReferenceObject }
  parameters?: { [parameter: string]: ParameterObject | ReferenceObject }
  examples?: { [example: string]: ExampleObject | ReferenceObject }
  requestBodies?: { [request: string]: RequestBodyObject | ReferenceObject }
  headers?: { [header: string]: HeaderObject | ReferenceObject }
  securitySchemes?: {
    [securityScheme: string]: SecuritySchemeObject | ReferenceObject
  }
  links?: { [link: string]: LinkObject | ReferenceObject }
  callbacks?: { [callback: string]: CallbackObject | ReferenceObject }
}

/**
 * Rename it to Paths Object to be consistent with the spec
 * See https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.0.md#pathsObject
 */
export interface PathsObject extends ISpecificationExtension {
  // [path: string]: PathItemObject;
  [path: string]: PathItemObject | any // Hack for allowing ISpecificationExtension
}

/**
 * @deprecated
 * Create a type alias for backward compatibility
 */
export type PathObject = PathsObject

export function getPath(
  pathsObject: PathsObject,
  path: string,
): PathItemObject | undefined {
  if (SpecificationExtension.isValidExtension(path)) {
    return undefined
  }
  return pathsObject[path] as PathItemObject
}

export interface PathItemObject extends ISpecificationExtension {
  $ref?: string
  summary?: string
  description?: string
  get?: OperationObject
  put?: OperationObject
  post?: OperationObject
  delete?: OperationObject
  options?: OperationObject
  head?: OperationObject
  patch?: OperationObject
  trace?: OperationObject
  servers?: ServerObject[]
  parameters?: Array<ParameterObject | ReferenceObject>
}
// ...
export interface ExternalDocumentationObject extends ISpecificationExtension {
  description?: string
  url: string
}
// ...
export interface ParameterObject extends BaseParameterObject {
  name: string
  in: ParameterLocation // "query" | "header" | "path" | "cookie";
}
export interface RequestBodyObject extends ISpecificationExtension {
  description?: string
  content: ContentObject
  required?: boolean
}
// ...
export interface ResponseObject extends ISpecificationExtension {
  description: string
  headers?: HeadersObject
  content?: ContentObject
  links?: LinksObject
}
export interface CallbacksObject extends ISpecificationExtension {
  // [name: string]: CallbackObject | ReferenceObject;
  [name: string]: CallbackObject | ReferenceObject | any // Hack for allowing ISpecificationExtension
}
export interface CallbackObject extends ISpecificationExtension {
  // [name: string]: PathItemObject;
  [name: string]: PathItemObject | any // Hack for allowing ISpecificationExtension
}
export interface HeadersObject {
  [name: string]: HeaderObject | ReferenceObject
}
export interface ExampleObject {
  summary?: string
  description?: string
  value?: any
  externalValue?: string
  [property: string]: any // Hack for allowing ISpecificationExtension
}
export interface LinksObject {
  [name: string]: LinkObject | ReferenceObject
}
export interface LinkObject extends ISpecificationExtension {
  operationRef?: string
  operationId?: string
  parameters?: LinkParametersObject
  requestBody?: any | string
  description?: string
  server?: ServerObject
  [property: string]: any // Hack for allowing ISpecificationExtension
}
export interface LinkParametersObject {
  [name: string]: any | string
}
export interface HeaderObject extends BaseParameterObject {
  $ref?: string
}
export interface TagObject extends ISpecificationExtension {
  name: string
  description?: string
  externalDocs?: ExternalDocumentationObject
  [extension: string]: any // Hack for allowing ISpecificationExtension
}
export interface ExamplesObject {
  [name: string]: ExampleObject | ReferenceObject
}

export interface ReferenceObject {
  $ref: string
}
// ...
export interface SchemaObject extends ISpecificationExtension {
  nullable?: boolean
  discriminator?: DiscriminatorObject
  readOnly?: boolean
  writeOnly?: boolean
  xml?: XmlObject
  externalDocs?: ExternalDocumentationObject
  example?: any
  examples?: any[]
  deprecated?: boolean

  type?: SchemaObjectType | SchemaObjectType[]
  format?: SchemaObjectFormat
  allOf?: Array<SchemaObject | ReferenceObject>
  oneOf?: Array<SchemaObject | ReferenceObject>
  anyOf?: Array<SchemaObject | ReferenceObject>
  not?: SchemaObject | ReferenceObject
  items?: SchemaObject | ReferenceObject
  properties?: { [propertyName: string]: SchemaObject | ReferenceObject }
  additionalProperties?: SchemaObject | ReferenceObject | boolean
  description?: string
  default?: any

  title?: string
  multipleOf?: number
  maximum?: number
  /** @desc OpenAPI 3.0: boolean, OpenAPI 3.1: number */
  exclusiveMaximum?: number | boolean
  minimum?: number
  /** @desc OpenAPI 3.0: boolean, OpenAPI 3.1: number */
  exclusiveMinimum?: number | boolean
  maxLength?: number
  minLength?: number
  pattern?: string
  maxItems?: number
  minItems?: number
  uniqueItems?: boolean
  maxProperties?: number
  minProperties?: number
  required?: string[]
  enum?: any[]
}
// ...
export interface SecuritySchemeObject extends ISpecificationExtension {
  type: SecuritySchemeType
  description?: string
  name?: string // required only for apiKey
  in?: string // required only for apiKey
  scheme?: string // required only for http
  bearerFormat?: string
  flows?: OAuthFlowsObject // required only for oauth2
  openIdConnectUrl?: string // required only for openIdConnect
}
// ...
export interface SecurityRequirementObject {
  [name: string]: string[]
}